Revert "Bug 964180 - Introduce AMD to settings app"
This reverts commit d9fb81e2d83dae21e53496b697cdd10ad6a2d4fd.

Conflicts:
	apps/communications/dialer/test/unit/multi_sim_action_button_test.js
	apps/settings/index.html
