--- apps/settings/.jshintrc
@@ -1,8 +1,0 @@
-{
-  "extends": "../../.jshintrc",
-  "predef": [
-    "define",
-    "require",
-    "Settings"
-  ]
-}
