+++ apps/settings/elements/carrier.html
@@ -79,7 +79,6 @@
     </div>
 
     <script src="shared/js/apn_helper.js"></script>
-    <script src="shared/js/settings_helper.js"></script>
     <script src="js/carrier.js"></script>
 
   </template>
