+++ apps/settings/index.html
@@ -66,7 +66,7 @@
     <script defer="" src="js/firefox_accounts/menu_loader.js"></script>
 
     <!-- Specific code -->
-    <script data-main="js/main" src="js/vendor/alameda.js"></script>
+    <script defer="" src="js/settings.js"></script>
 
     <!-- shared helper library -->
     <!--<script defer src="shared/js/download/download_store.js"></script>-->
