+++ apps/settings/test/unit/.jshintrc
@@ -10,7 +10,6 @@
     "suiteTeardown",
     "require",
     "requireApp",
-    "testRequire",
     "sinon",
     "mocha",
     "loadBodyHTML"
