--- apps/settings/test/unit/mock_page_transitions.js
@@ -1,12 +1,0 @@
-/* global define */
-define(function() {
-  'use strict';
-  return {
-    oneColumn: function(oldPanel, newPanel, callback) {
-      setTimeout(callback);
-    },
-    twoColumn: function(oldPanel, newPanel, callback) {
-      setTimeout(callback);
-    }
-  };
-});
