+++ apps/system/test/unit/keyboard_helper_test.js
@@ -122,6 +122,7 @@ suite('KeyboardHelper', function() {
     navigator.mozApps = MockNavigatormozApps;
 
     realMozSettings = navigator.mozSettings;
+    MockNavigatorSettings.mSyncRepliesOnly = true;
     navigator.mozSettings = MockNavigatorSettings;
 
     // ensure the default settings are indeed default
@@ -130,12 +131,12 @@ suite('KeyboardHelper', function() {
   });
 
   suiteTeardown(function() {
+    MockNavigatorSettings.mSyncRepliesOnly = false;
     navigator.mozApps = realMozApps;
     navigator.mozSettings = realMozSettings;
   });
 
   setup(function() {
-    MockNavigatorSettings.mSyncRepliesOnly = true;
     // reset KeyboardHelper each time
     KeyboardHelper.settings.enabled = {};
     KeyboardHelper.settings['default'] = defaultSettings['default'];
