+++ shared/js/settings_listener.js
@@ -60,3 +60,4 @@ var SettingsListener = {
     });
   }
 };
+
